requirements.txt
prettytable==2.1.0
Pyrebase4==4.4.3
timeago==1.0.15
python-dotenv